backend/RAG_python-quiz/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.9

psycopg2-binary==2.9.9
bcrypt>=4.2.0
python-jose[cryptography]==3.3.0

requests==2.32.3
openai>=1.0.0
google-genai>=1.0.0

numpy<2.0.0
pandas>=2.0.0
matplotlib>=3.8.0
reportlab>=4.0.0
pypdf==4.3.1

langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2
langchain-openai==0.1.25
langgraph==0.2.0
markdownify>=1.2.0

ragas>=0.4.0
rich>=13.0.0
